gaus update

+++ gaussFit_goffsetPG/gaussFit_goffsetPG_v1.0/gaussFit_main_goffsetPG_v1.0.bash
@@ -19,12 +19,12 @@ noise_ref=( 5.8 6.0 )
 PHA0_ref=( 958 )
 grade_ref=( 0 2 3 4 6 )
 
+# spec dir
+d_dat='dat'
+
 # args
 specType=$1
 phType=$2
-
-# dir
-d_dat='dat'
 
 # file
 b_fit1='gaussFit_sub_goffsetPG_v1.0.bash'
@@ -38,12 +38,14 @@ if [ "${specType}" = "org" ] && [ "${phType}" = "PHA" ] ; then
 elif [ "${specType}" = "org" ] && [ "${phType}" = "PHAS_SUM" ] ; then
     fileType="orgSpec_sum"
     d_output="fittingResult_org_sum"
+    grade_ref+=8; grade_ref+=9
 elif [ "${specType}" = "cor" ] && [ "${phType}" = "PHA" ] ; then
     fileType="corSpec"
     d_output="fittingResult_cor"
 elif [ "${specType}" = "cor" ] && [ "${phType}" = "PHAS_SUM" ] ; then
     fileType="corSpec_sum"
     d_output="fittingResult_cor_sum"
+    grade_ref+=8; grade_ref+=9
 else
     echo "incollect args"
     exit
@@ -76,8 +78,17 @@ for PHA0 in ${PHA0_ref[@]} ; do
         
         for grade in ${grade_ref[@]} ; do
             
+            # read file
+            readFile="${d_dat}/${fileType}_PHA${PHA0}_noise${noise}_G${grade}.dat"
+            
+            # check file
+            if [ ! -e $readFile ] ; then
+                echo "${readFile} does not exist"
+                continue
+            fi
+            
             # fit
-            bash $b_fit1 ${d_dat}/${fileType}_PHA${PHA0}_noise${noise}_G${grade}.dat
+            bash $b_fit1 $readFile
             
             # rename
             mv $f_fitData fitData_${fileType}_PHA${PHA0}_noise${noise}_G${grade}.dat
